nu: initial efforts for handling internal CPU congestion.

+++ inc/nu/commons.hpp
@@ -23,8 +23,8 @@ struct ErasedType {};
 struct ProcletHeader;
 
 struct Resource {
-  uint32_t cores;
-  uint32_t mem_mbs;
+  float cores;
+  float mem_mbs;
 };
 
 struct VAddrRange {
